Rewrite UserTurn and ComputerTurn. Add while loops to main to allow game to run
This is basically a complete rewrite of the logic of how the game works. Before, the functions for the computers turn and the users turn called each other. Now, the turn functions return the total count of sticks remaining. The looping logic now exists in main using the returned value from the two turn functions.